Fix bug in DynamoDB global table support when deploying multiple times
If you had a global table defined on a table that used provisioned
capacity billing mode, and you deployed that table multiple times, on
the second (and all subsequent) deploys, you'd get this error:

> An error occurred: GlobalTableResourceName - Failed to update
resource. The provisioned throughput for the table will not change. The
requested value equals the current value. Current ReadCapacityUnits
provisioned for the table: 2. Requested ReadCapacityUnits: 2. Current
WriteCapacityUnits provisioned for the table: 2. Requested
WriteCapacityUnits: 2. Refer to the Amazon DynamoDB Developer Guide for
current limits and how to request higher limits..

The problem was that we were trying to set the provisioned throughput
even though there was no change. We were doing that because of faulty
logic introduced in a previous commit because we _must_ supply the
provisioned throughput during a change from on-demand to provisioned,
but we must _not_ provide it when your billing mode was and is staying
as provisioned.