Stop managing DynamoDB tables in our own custom resource
The custom resource for creating a DynamoDB global table was excessively
complex. It allowed you to deploy DynamoDB tables in a single region,
and then when you used the custom resource to create the global table,
the resource would copy those tables to other regions for you before
creating the global table. This caused problems whenever new
CloudFormation support was added for new features of DynamoDB tables; it
was also needlessly complex on its own because of the complexity of
eventually-consistent descriptions of tables, etc.

Rather than continuing to manage DynamoDB tables ourselves, we instead
now recommend creating the (empty) tables in all the regions you want
them in using CloudFormation natively. Then (perhaps in a different
stack), use the custom resource to create the global table on its own.
This drastically reduces the complexity of the custom resource.

For one single RC release (1.0.0-rc3), we will leave the old custom
resource in the codebase, working exactly as it always has. If you want
to use the newer, simpler type of global table resource, add a
`IsSimpleType: true` property to your global table resource. Then, in
the following RC release (rc4), the old "complex" global table will go
away and the newer "simple" resource will remain, without the need for
the `IsSimpleType` property.