fix: only set BillingMode when it differs between src and dest
Updating a GSI on a global table using provisioned capacity was failing with `One or more
parameter values were invalid: ProvisionedThroughput must be specified when BillingMode is
PROVISIONED`. This was due to BillingMode getting set on the table, but not
ProvisionedThroughput. This commit fixes this as it makes the custom resource only set
BillingMode when it actually has changed.